+++ src/integer64.c
@@ -680,10 +680,12 @@ SEXP min_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
   long long i, n = LENGTH(e1_);
   long long * e1 = (long long *) REAL(e1_);
   long long * ret = (long long *) REAL(ret_);
+  bool onlyNas = true;
   ret[0] = MAX_INTEGER64;
     if (asLogical(na_rm_)){
         for(i=0; i<n; i++){
             if (e1[i]!=NA_INTEGER64 && e1[i]<ret[0]){
+                onlyNas = false;       
                 ret[0] = e1[i];
             }
         }
@@ -693,11 +695,15 @@ SEXP min_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
                 ret[0] = NA_INTEGER64;
                 return ret_;
             }else{
+                onlyNas = false;       
                 if (e1[i]<ret[0])
                     ret[0] = e1[i];
             }
         }
     }
+  if (onlyNas){
+    ret[0] = NA_INTEGER64;
+  }
   return ret_;
 }
 
@@ -705,10 +711,12 @@ SEXP max_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
   long long i, n = LENGTH(e1_);
   long long * e1 = (long long *) REAL(e1_);
   long long * ret = (long long *) REAL(ret_);
+  bool onlyNas = true;
   ret[0] = MIN_INTEGER64;
     if (asLogical(na_rm_)){
         for(i=0; i<n; i++){
             if (e1[i]!=NA_INTEGER64 && e1[i]>ret[0]){
+                onlyNas = false;       
                 ret[0] = e1[i];
             }
         }
@@ -718,11 +726,15 @@ SEXP max_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
                 ret[0] = NA_INTEGER64;
                 return ret_;
             }else{
+                onlyNas = false;       
                 if (e1[i]>ret[0])
                     ret[0] = e1[i];
             }
         }
     }
+  if (onlyNas){
+    ret[0] = NA_INTEGER64;
+  }
   return ret_;
 }
 
@@ -730,11 +742,13 @@ SEXP range_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
   long long i, n = LENGTH(e1_);
   long long * e1 = (long long *) REAL(e1_);
   long long * ret = (long long *) REAL(ret_);
+  bool onlyNas = true;
   ret[0] = MAX_INTEGER64;
   ret[1] = MIN_INTEGER64;
     if (asLogical(na_rm_)){
         for(i=0; i<n; i++){
             if (e1[i]!=NA_INTEGER64){
+                onlyNas = false;
                 if (e1[i]<ret[0])
                     ret[0] = e1[i];
                 if (e1[i]>ret[1])
@@ -747,6 +761,7 @@ SEXP range_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
                 ret[0] = ret[1] = NA_INTEGER64;
                 return ret_;
             }else{
+                onlyNas = false;
                 if (e1[i]<ret[0])
                     ret[0] = e1[i];
                 if (e1[i]>ret[1])
@@ -754,6 +769,10 @@ SEXP range_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
             }
         }
     }
+  if (onlyNas){
+    ret[0] = NA_INTEGER64;
+    ret[1] = NA_INTEGER64;
+  }
   return ret_;
 }